microservice/backend/doctor_notification_service/requirements.txt
# requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
httpx==0.23.3
apscheduler==3.9.1
python-dotenv==1.0.0
pydantic==1.10.7
python-jose[cryptography]
PyJWT==2.8.0
redis~=6.2.0
aio-pika==9.3.0
pytest
pytest-asyncio